+++ projectApp/src/components/LocalCSVReader.tsx
@@ -2,8 +2,7 @@ import * as fsPromise from 'fs/promises';
 import * as fs from 'fs';
 import Papa from 'papaparse';
 
-import CSVHeaders from '../data_structures/CSVHeaders';
-import TimeSeriesData from '../data_structures/TimeSeriesData';
+import {CSVHeaders, TimeSeriesData} from '../data_structures/CSVInterfaces';
 
 /**  
 * This function reads the headers of a csv file and stores it
