+++ projectApp/src/components/UrlCSVReader.tsx
@@ -1,7 +1,6 @@
 import Papa from 'papaparse';
 
-import CSVHeaders from '../data_structures/CSVHeaders';
-import TimeSeriesData from '../data_structures/TimeSeriesData';
+import { CSVHeaders, TimeSeriesData} from '../data_structures/CSVInterfaces';
 
 /**
  * Get the headers of a file at a url
