--- projectApp/src/data_structures/CSVHeaders.tsx
@@ -1,3 +1,0 @@
-export default interface CSVHeaders {
-    headers: string[];
-}
