+++ projectApp/src/data_structures/CSVInterfaces.tsx
@@ -1,0 +1,10 @@
+/*
+* Interfaces required for the CSV Readers
+*/
+export interface CSVHeaders {
+    headers: string[];
+}
+
+export interface TimeSeriesData {
+    [key: string]: string | number;
+}
