--- projectApp/src/data_structures/TimeSeriesData.tsx
@@ -1,3 +1,0 @@
-export default interface TimeSeriesData {
-    [key: string]: string | number;
-}
